pyproject.toml
[project]
name = "kasperjunge-blog"
version = "0.1.0"
description = "Simple blog builder"
requires-python = ">=3.10"
dependencies = [
    "markdown>=3.5",
    "python-frontmatter>=1.1",
]

[project.scripts]
build = "build:main"
